requirement.txt
asgiref==3.7.2
certifi==2023.7.22
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
Django==4.2.6
djangorestframework==3.14.0
fonttools==4.45.1
idna==3.4
joblib==1.3.2
kiwisolver==1.4.5
matplotlib==3.8.2
mysqlclient==2.2.0
numpy==1.23.5
packaging==23.2
pandas==1.5.3
Pillow==10.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0
scikit-learn==1.2.2
scipy==1.11.4
six==1.16.0
sqlparse==0.4.4
threadpoolctl==3.2.0
tzdata==2023.3
urllib3==2.1.0



Django==4.2.6
djangorestframework==3.14.0
joblib==1.3.2
mysqlclient==2.2.0
numpy==1.23.5
pandas==1.5.3
requests==2.31.0
scikit-learn==1.2.2
scipy==1.11.4
sqlparse==0.4.4